+++ pages/standards.tsx
@@ -3,7 +3,7 @@ import Container from "react-bootstrap/Container";
 import Row from "react-bootstrap/Row";
 import Breadcrumb from "react-bootstrap/Breadcrumb";
 import HtanNavbar from "../components/HtanNavbar";
-import {getContent} from "../util";
+import {getContent} from "../api_util";
 
 const Standards = () => {
     const content = getContent("data-standards","summary-blurb");
